Set OS/2-specific configurations
1. Set INSTALL to ginstall
A default value, install, conflicts with INSTALL file.

2. Set NO_TRUSTABLE_FILEMODE
On OS/2, the executable bit is not trustable.
This causes core.filemode to be set to false by default.

3. Set NO_POSIX_GOODIES
Some functions such as setsid() do not work.
Especially, this prevents git from dying due to setsid() failure when
daemonizing.

4. Set CSPRNG_METHOD to openssl
This implements csprng_bytes() with RAND_bytes() of openssl.

    modified:   config.mak.uname