Realloc environmental variables at startup on OS/2
This fixes the crash when trying to fixup commit to non-last commit
with git rebase -i command.

For example, consider this:

    pick commit1
    pick commit2
    pick commit3
    pick commit4

then, fixup like this:

    pick commit1
    pick commit2
    f    commit4
    pick commit3

In this case, git crashes due to SIGSEGV when applying the commit3 after
performing fixup commit 4 to commit 2.

The cause of SIGSEGV is to try to read from uncommited memory. The
memory is a prt of the environmental segment.

kLIBC seems to have a problem to copy the environmental segment to the
fork()ed child process.