Fix failure to execute git when a directory `git' exists
If a directory named is `git' exists in a current directory, `git'
command such as `git push' fails because the directory `git' is
searched first before the executable git.

This is a bug of kLIBC v0.6.6 which is already fixed, but the fixed
kLIBC has not been released, yet.

This is the regression of commit 45b74c.

    modified:   compat/os2.c