Fix failure to execute a script
If there is a directory in a current directory whose name is the same
as an interpreter name of a script, kLIBC v0.6.6 cannot execute it.

For example, git-add--interactive starts with #!perl. By the way, if
a directory named perl exists in a current directory, `git add -p'
fails because a directory perl is executed not the executable perl.

To execute such scripts correctly, use sh -c.

Fixed #1.

    modified:   compat/os2.c