Use a socket as a pipe by default
kLIBC programs support a socket as a pipe, so supporting this is a good
choice.

Especially, newer ssh.exe using kLIBC expects a socket not a real pipe.

And internally, kLIBC select() does not support a socket at all. This
situation can be improved by this.

You can change this by setting GIT_USE_NPIPE to the following values:

1. 'S' or 's': socket
2. 'T', 't', 'Y', 'y', 'O', 'o', any other values than 0: named pipe
3. anything else: unnamed pipe

    modified:   compat/os2.c