Fix [git-gc] destroy a repository
When socketpair() is used as a back-end for pipe(), [git-gc] destructs
a repository at all.

This is because [git-repack] cannot read anything from cmd.out. Reading
from cmd.out does not wait for it to be filled even if it is in blocking
mode if fdout[1] is closed.

As a result, old pack files are removed but new temporary pack files are
not renamed. And they are removed. This leads to the destruction of a
repository.

However, waiting about 500ms before closing fdout[1] could avoid this
problem.

Although I don't know the exact reason, my guess based on the above is
that fdout[1] is closed before kLIBC completes to passes stdios which
are sockets to a child process. After all, all the handles are closed
before writing to them, and reading from an opposite end fails.

This patch should be removed if the root cause of socketpair() problem
is resolved.